# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * DevDex Game Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface User {
  id: number;
  username: string;
  email: string;
  /** @nullable */
  avatarUrl: string | null;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
}

export interface RegisterInput {
  /** @minLength 3 */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface Game {
  id: number;
  title: string;
  /** @nullable */
  description: string | null;
  gameUrl: string;
  /** @nullable */
  coverImageUrl: string | null;
  slug: string;
  /** @nullable */
  category: string | null;
  featured: boolean;
  playCount: number;
  authorId: number;
  author: User;
  createdAt: string;
  updatedAt: string;
}

export interface GameInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  gameUrl: string;
  coverImageUrl?: string;
  category?: string;
}

export interface GameUpdate {
  title?: string;
  description?: string;
  gameUrl?: string;
  coverImageUrl?: string;
  category?: string;
  featured?: boolean;
}

export interface GameListResponse {
  games: Game[];
  total: number;
}

export interface PlatformStats {
  totalGames: number;
  totalUsers: number;
  totalPlays: number;
}

export interface Group {
  id: number;
  name: string;
  /** @nullable */
  description: string | null;
  slug: string;
  /** @nullable */
  coverImageUrl: string | null;
  isPublic: boolean;
  authorId: number;
  memberCount: number;
  author: User;
  createdAt: string;
  updatedAt: string;
}

export interface GroupMember {
  id: number;
  userId: number;
  groupId: number;
  role: string;
  joinedAt: string;
  user: User;
}

export interface GroupDetail {
  id: number;
  name: string;
  /** @nullable */
  description: string | null;
  slug: string;
  /** @nullable */
  coverImageUrl: string | null;
  isPublic: boolean;
  authorId: number;
  memberCount: number;
  author: User;
  members: GroupMember[];
  isMember: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface GroupInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  coverImageUrl?: string;
  isPublic?: boolean;
}

export interface GroupListResponse {
  groups: Group[];
  total: number;
}

export type ListGamesParams = {
search?: string;
limit?: number;
offset?: number;
};

export type SearchGamesParams = {
q: string;
};

export type ListGroupsParams = {
limit?: number;
offset?: number;
};