# lib/api-zod/src/generated/types/game.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * DevDex Game Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface Game {
  id: number;
  title: string;
  /** @nullable */
  description: string | null;
  gameUrl: string;
  /** @nullable */
  coverImageUrl: string | null;
  slug: string;
  /** @nullable */
  category: string | null;
  featured: boolean;
  playCount: number;
  authorId: number;
  author: User;
  createdAt: string;
  updatedAt: string;
}