# lib/api-zod/src/generated/types/gameListResponse.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * DevDex Game Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Game } from './game';

export interface GameListResponse {
  games: Game[];
  total: number;
}